thrp: split all classes in separate files
First of all, having all the UserParser implementations in one place
was a bad idea from the get-go: I kept catching myself editing the
wrong class many times, hence the split.

TouhouCryptFile was split from TouhouReplay because, in the future,
th06 and th07 are going to be parsed through USER parser, which would
require accessing CryptFile from outside, so it's declaration would
need to be moved into it's own hpp file.

TouhouUserReader was split for the same reason, pretty much.

wrt ITouhouUserParser + TouhouUserParserFactory, idk, I just did it
impulsively. :^) I'll merge those classes together in the next commit
anyway.

p.s. the reason why putting multiple classes in a header is bad, is
because when you change one of them, all the depending translation
units will have to recompile, even if they don't use the class that
was changed.